requirements.txt
discord.py==1.7.3
discordbot.py==0.2.3a3

const lib = require('lib')({token: process.env.STDLIB_SECRET_TOKEN});

await lib.discord.channels['@0.3.0'].messages.create({
  "channel_id": `${context.params.event.channel_id}`,
  "content": "",
  "tts": false,
  "embeds": [
    {
      "type": "rich",
      "title": `[bot]`,
      "description": `requests failed`,
      "color": 0x00FFFF,
      "fields": [
        {
          "name": `asdfsafasfd`,
          "value": "\u200B",
          "inline": true
        },
        {
          "name": `sfasdf`,
          "value": "\u200B"
        }
      ],
      "url": `https://www.google.com`
    }
  ]
});